requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
absl-py=0.10.0=py37_0
aiohttp=3.6.2=pypi_0
albumentations=1.2.1=pypi_0
apex=0.1=pypi_0
argon2-cffi=20.1.0=py37he774522_1
astor=0.8.1=py37_0
async-timeout=3.0.1=pypi_0
async_generator=1.10=py37h28b3542_0
atari-py=0.2.9=pypi_0
attrs=20.2.0=py_0
av=8.0.2=pypi_0
backcall=0.2.0=py_0
blas=1.0=mkl
bleach=3.2.1=py_0
blinker=1.4=py37_0
brotlipy=0.7.0=py37he774522_1000
ca-certificates=2022.4.26=haa95532_0
cachetools=4.1.1=py_0
certifi=2022.6.15=py37haa95532_0
cffi=1.14.3=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
cloudpickle=2.2.0=pypi_0
colorama=0.4.4=py_0
crc32c=2.2.post0=pypi_0
cryptography=3.1.1=py37h7a1dbc1_0
cudatoolkit=11.0.221=h74a9793_0
cvxpy=1.1.17=pypi_0
cycler=0.10.0=py37_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
dm-tree=0.1.7=pypi_0
eagerpy=0.27.0=pypi_0
ecos=2.0.7.post1=pypi_0
entrypoints=0.3=py37_0
facenet-pytorch=2.5.0=pypi_0
faiss=1.7.2=pypi_0
faiss-cpu=1.6.5=pypi_0
foolbox=3.0.1=pypi_0
freetype=2.10.4=hd328e21_0
future=0.18.2=pypi_0
fuzzy-c-means=1.6.3=pypi_0
fxpmath=0.4.0=pypi_0
gast=0.2.2=pypi_0
gitdb=4.0.5=pypi_0
gitpython=3.1.11=pypi_0
google-auth=1.22.0=pypi_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
grpcio=1.32.0=pypi_0
gym=0.9.2=pypi_0
h5py=2.8.0=py37hf7173ca_0
hdf5=1.8.18=vc14h7a021fe_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=vc14hc45fdbb_0
idna=2.10=py_0
imageio=2.21.1=pypi_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2019.4=245
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.18.1=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_1
jedi=0.17.2=py37_0
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9b=vc14h4d7706e_1
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.6.3=py37_0
jupyterlab=2.2.8=py_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=1.2.0=py_0
keras=2.3.1=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.2.0=py37h74a9793_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libfaiss=1.7.2=h2e52968_0_cpu
libmklml=2019.0.5=haa95532_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.19.1=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
llvmlite=0.34.0=py37h1a82afc_4
lz4-c=1.9.2=hf4a77e7_3
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.2.2=pypi_0
markupsafe=1.1.1=py37hfa6e2cd_1
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py37hba9282a_0
mistune=0.8.4=py37hfa6e2cd_1001
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.0=py37h675688f_0
mmdnn=0.3.1=pypi_0
msys2-conda-epoch=20160418=1
multidict=4.7.6=py37he774522_1
nbclient=0.5.1=py_0
nbconvert=6.0.7=py37_0
nbformat=5.0.8=py_0
nest-asyncio=1.4.1=py_0
networkx=2.6.3=pypi_0
ninja=1.10.1=py37h7ef1ec2_0
notebook=6.1.4=py37_0
numba=0.51.2=py37h414f9d2_0
numpy=1.21.4=pypi_0
numpy-base=1.19.1=py37ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py37_0
onnx=1.8.1=pypi_0
onnx-tf=1.8.0=pypi_0
onnx2keras=0.0.24=pypi_0
opencv-contrib-python=4.5.1.48=pypi_0
opencv-python=4.5.5.64=pypi_0
opencv-python-headless=4.6.0.66=pypi_0
openssl=1.1.1o=h2bbff1b_0
opt_einsum=3.1.0=py_0
osqp=0.6.2.post0=pypi_0
packaging=20.4=py_0
pandas=1.1.1=py37ha925a31_0
pandoc=2.11=h9490d1a_0
pandocfilters=1.4.2=py37_1
parso=0.7.0=py_0
pickleshare=0.7.5=py37_1001
pillow=8.0.0=py37hca74424_0
pip=20.2.4=py37_0
plotly=5.6.0=pypi_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.17.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pychemometrics=0.13.6=pypi_0
pycparser=2.20=py_2
pydantic=1.8.2=pypi_0
pyglet=1.5.17=pypi_0
pygments=2.7.1=py_0
pyjwt=1.7.1=py37_0
pynndescent=0.5.6=pyh6c4a22f_0
pyopengl=3.1.5=pypi_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37ha878b3d_0
pyrsistent=0.17.3=py37he774522_0
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda110_cudnn8_0
pytorch2keras=0.2.4=pypi_0
pytz=2020.1=py_0
pywavelets=1.3.0=pypi_0
pywin32=227=py37he774522_1
pywinpty=0.5.7=py37_0
pyyaml=5.3.1=pypi_0
pyzmq=19.0.2=py37ha925a31_1
qdldl=0.1.5.post0=pypi_0
qt=5.9.7=vc14h73c81de_0
qtconsole=4.7.7=py_0
qtpy=1.9.0=py_0
qudida=0.0.4=pypi_0
requests=2.24.0=py_0
requests-oauthlib=1.3.0=py_0
rsa=4.6=py_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0=py37ha78be43_1
scipy=1.7.3=pypi_0
scs=2.1.4=pypi_0
seaborn=0.11.0=py_0
send2trash=1.5.0=py37_0
setuptools=50.3.0=py37h9490d1a_1
sip=4.19.24=py37ha925a31_0
six=1.15.0=py_0
smmap=3.0.4=pypi_0
sqlite=3.38.5=h2bbff1b_0
tabulate=0.8.9=pypi_0
tbb=2021.5.0=h2d74725_0
tenacity=8.0.1=pypi_0
tensorboard=2.3.0=pypi_0
tensorboard-data-server=0.6.0=py37haa95532_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorflow=2.1.0=mkl_py37ha977152_0
tensorflow-addons=0.18.0=pypi_0
tensorflow-base=2.1.0=mkl_py37h230818c_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
tensorflow-probability=0.18.0=pypi_0
termcolor=1.1.0=py37_1
terminado=0.9.1=py37_0
testpath=0.4.4=py_0
tfrecord=1.14.1=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.11.2=pypi_0
timm=0.6.7=pypi_0
tk=8.6.7=vc14hb68737d_1
torch=1.6.0=pypi_0
torchaudio=0.7.2=py37
torchmetrics=0.9.3=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.8.2=py37_cu110
tornado=6.0.4=py37he774522_1
tqdm=4.49.0=py_0
traitlets=5.0.5=py_0
typeguard=2.13.3=pypi_0
typer=0.3.2=pypi_0
typing_extensions=3.7.4.3=py_0
umap-learn=0.5.2=py37h03978a9_1
urllib3=1.25.11=py_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.1=pypi_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py37_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
winpty=0.4.3=4
wrapt=1.12.1=py37he774522_1
xz=5.2.5=h62dcd97_0
yarl=1.6.0=pypi_0
zeromq=4.3.2=ha925a31_3
zipp=3.3.1=py_0
zlib=1.2.11=vc14h1cdd9ab_1
zstd=1.4.4=ha9fde0e_3
